폴더 store, action, saga 구조로 변경
