+++ client/src/reducer/index.js
@@ -1,0 +1,6 @@
+import { combineReducers } from "redux";
+import todoReducer from './todoReducer';
+
+export default combineReducers({
+	todoReducer,
+});
