+++ client/src/sagas/index.js
@@ -1,7 +1,7 @@
-// import { all, fork } from 'redux-saga/effects';
-// import todoWatcher from './todoSaga';
-// export default function* rootSaga() {
-//     yield all([
-//         fork(todoWatcher),
-//     ])
-// }
+import { all, fork } from 'redux-saga/effects';
+import todoWatcher from './todoSaga';
+export default function* rootSaga() {
+    yield all([
+        fork(todoWatcher),
+    ])
+}
