--- client/src/store/modules/index.js
@@ -1,7 +1,0 @@
-import { combineReducers } from 'redux';
-import todoAction from './todoAction';
-
-export default combineReducers({
-	todoAction,
-	// 다른 리듀서를 만들게되면 여기에 넣어줌..
-});
